update on the notifications

+++ app/Http/Controllers/UserController.php
@@ -50,9 +50,16 @@ class UserController extends Controller
      */
     public function show($id)
     {
-        $user=User::find($id);
-     return new UserResource($user);
+        if(User::find($id))
+        {
+             $user=User::find($id);
+            return new UserResource($user);
         }
+        else{
+            return "There is no user with this id";
+        }
+       
+    }
 
     /**
      * Update the specified resource in storage.
@@ -63,16 +70,21 @@ class UserController extends Controller
      */
     public function update(Request $request, $id)
     {
+        if(User::find($id))
+        {
         $user = User::find($id);
         $user->name=$request->name;
         $user->email=$request->email;
         $user->phone=$request->phone;
         $user->address=$request->address;
         // $user->password=$request->password;
-
+        
         $user->save();
         Mail::to($request->email)->send(new UpdatedUserMail());
         return "Updated";
+        }
+        else
+        return "There is no user with this id";
     }
 
     /**
@@ -83,8 +95,14 @@ class UserController extends Controller
      */
     public function destroy($id)
     {
-        User::destroy($id);
-        Mail::to($request->email)->send(new DeleteUserMail());
-        return "destroyed";
+        if(User::find($id)){
+            $user = User::find($id)->email;
+            User::destroy($id);
+            Mail::to($request->email)->send(new DeleteUserMail());
+            return "destroyed";
+        }
+        else
+        return "There is no user with this id";
+       
     }
 }
